+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.7.0 - 2024/01/26
+
+### Changes
+
+- Text Completion using Chat Completion at the backend with default model as GPT 3.5 Turbo
+- Added property usage - total token usage to Chat Bot and Text Completion.
+
+## v0.6.0 - 2024/01/22
+
+### Changes
+
+- Migrated from Betalgo to Azure Open AI 1.0.0-beta.12
+- Text Completion default model updated.
+
+## v0.5.0 - 2024/01/03
+
+### Changes
+
+- Moved to Microsoft repository and references updated
+
 ## v0.4.0 - 2023/11/14
 
 ### Added